Moved AppUrl to common module and fix pixel service and entity
