+++ statistics/src/main/java/com/duckduckgo/app/statistics/model/PixelEntity.kt
@@ -19,8 +19,8 @@ package com.duckduckgo.app.statistics.model
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import androidx.room.TypeConverter
-import com.duckduckgo.app.di.JsonModule
 import com.squareup.moshi.JsonAdapter
+import com.squareup.moshi.Moshi
 import com.squareup.moshi.Types
 
 @Entity(tableName = "pixel_store")
@@ -46,7 +46,7 @@ class QueryParamsTypeConverter {
 
 class Adapters {
     companion object {
-        private val moshi = JsonModule().moshi()
+        private val moshi = Moshi.Builder().build()
         private val mapStringStringType = Types.newParameterizedType(Map::class.java, String::class.java, String::class.java)
         val queryParamsAdapter: JsonAdapter<Map<String, String>> = moshi.adapter(mapStringStringType)
     }
